Move ColorFilter component and make it use props passed from parent
